+++ common/src/commonMain/kotlin/com/projeto/nutrilog/di/Koin.kt
@@ -8,13 +8,14 @@ import com.projeto.nutrilog.database.NutriLogDatabase
 import app.cash.sqldelight.db.SqlDriver
 import com.projeto.nutrilog.domain.repository.UserRepository
 import com.projeto.nutrilog.data.repository.UserRepositoryImpl
-import com.projeto.nutrilog.domain.repository.ProgressRepository
-import com.projeto.nutrilog.data.repository.ProgressRepositoryImpl
+import com.projeto.nutrilog.domain.repository.MealLogRepository
+import com.projeto.nutrilog.data.repository.MealLogRepositoryImpl
 import com.projeto.nutrilog.domain.usecase.GetUserUseCase
 import com.projeto.nutrilog.domain.usecase.SaveUserUseCase
 import com.projeto.nutrilog.domain.usecase.ClearUserUseCase
-import com.projeto.nutrilog.domain.usecase.GetProgressUseCase
-import com.projeto.nutrilog.domain.usecase.AddConsumptionUseCase
+import com.projeto.nutrilog.domain.usecase.GetMealLogsUseCase
+import com.projeto.nutrilog.domain.usecase.AddMealLogUseCase
+import com.projeto.nutrilog.domain.usecase.DeleteMealLogUseCase
 import com.projeto.nutrilog.presentation.register.RegisterViewModel
 import com.projeto.nutrilog.presentation.main.MainViewModel
 import com.projeto.nutrilog.presentation.dashboard.DashboardViewModel
@@ -36,13 +37,14 @@ fun commonModule() = module {
         NutriLogDatabase(driver)
     }
     single<UserRepository> { UserRepositoryImpl(get()) }
-    single<ProgressRepository> { ProgressRepositoryImpl(get()) }
+    single<MealLogRepository> { MealLogRepositoryImpl(get()) }
     single { GetUserUseCase(get()) }
     single { SaveUserUseCase(get()) }
     single { ClearUserUseCase(get(), get()) }
-    single { GetProgressUseCase(get()) }
-    single { AddConsumptionUseCase(get()) }
+    single { GetMealLogsUseCase(get()) }
+    single { AddMealLogUseCase(get()) }
+    single { DeleteMealLogUseCase(get()) }
     viewModel { RegisterViewModel(get()) }
     viewModel { MainViewModel(get(), get()) }
-    viewModel { DashboardViewModel(get(), get(), get()) }
+    viewModel { DashboardViewModel(get(), get(), get(), get()) }
 }