+++ common/src/commonMain/kotlin/com/projeto/nutrilog/domain/usecase/DeleteMealLogUseCase.kt
@@ -1,0 +1,9 @@
+package com.projeto.nutrilog.domain.usecase
+
+import com.projeto.nutrilog.domain.repository.MealLogRepository
+
+class DeleteMealLogUseCase(private val repository: MealLogRepository) {
+    suspend operator fun invoke(id: Long) {
+        repository.deleteMealLog(id)
+    }
+}